Some deletions, imports cleanup and code refactor

+++ src/main/java/org/ginafro/notenoughfakepixel/utils/StringUtils.java
@@ -9,29 +9,24 @@ import java.util.NavigableMap;
 
 public class StringUtils {
 
-
     public static int romanToNumerical(String s){
         int total = 0;
         for(int i =0;i < s.length();i++){
             char r = s.charAt(i);
             int s1 = value(r);
-            if (i+1 <s.length())
-            {
+            if (i+1 <s.length()) {
                 int s2 = value(s.charAt(i+1));
-//comparing the current character from its right character
-                if (s1 >= s2)
-                {
-//if the value of current character is greater or equal to the next symbol
+                //comparing the current character from its right character
+                if (s1 >= s2) {
+                    //if the value of current character is greater or equal to the next symbol
                     total = total + s1;
                 }
-                else
-                {
-//if the value of the current character is less than the next symbol
+                else {
+                    //if the value of the current character is less than the next symbol
                     total = total - s1;
                 }
             }
-            else
-            {
+            else {
                 total = total + s1;
             }
         }
@@ -107,7 +102,7 @@ public class StringUtils {
             }
         }
 
-        double d = ((long) n / 100) / 10.0;
+        double d = ((double) (long) n / 100) / 10.0;
         boolean isRound = (d * 10) % 10 == 0;
         return d < 1000
                 ? (isRound || d > 9.99 ? (int) d * 10 / 10 : d + "") + "" + sizeSuffix[iteration]
